fix(#45): catch RPC call req builder error on invalid URI
The examples in the docs were misleading, so it was easy to configure
invalid settings via env file for running it in docker.

To make matters worse, we were crashing on an uncaught error when
the endpoint variable was malformed (contained double quotes) and made
it all the way to the collector.

This PR also contains docs improvements and consistency updates
around what users ought to be using for environment variables
and environment files.